Update rackspace-cloud-backup-install-the-agent-on-linux.md
https://trello.com/c/bSkhtoco
Request to update from Aaron Davis. Emailed Brett Johnson on backup agent team about Aaron's request.
Brett suggested renaming the article at https://www.rackspace.com/knowledge_center/article/update-the-rackspace-cloud-backup-agent to "Manually update..." and providing a link to that article at the top of this one - long with information to go there if the Cloub Backup updater fails. Also the title of this article was changed to "Install or update..."