add horizontal line and make minor edits
I viewed this topic when editing a new article about IP addresses. I noticed that the new "IP" section did not have a horizontal line above it, so I added one. I also removed the hyphen from the article title (not recommended), and removed the intro sentence from the new section (not needed).