rm conflicting information re crossdomain.xml mod
Please refer to: https://support.rackspace.com/how-to/cloud-files-faq/ --> Account Services section. 
Cloud Files CDN does not support exposing a custom crossdomain.xml file, as this is a required file by the OpenStack Swift project. OpenStack Swift uses this as a global configuration file for the installation, and can not be modified for multiple tenants, such as our Public Cloud. Creating a crossdomain.xml file in the root of a container will not make it available for use via Cloud Files CDN. The default crossdomain.xml file is returned by origin instead.