Update adding-a-rackspace-email-group-list.md
Updating Item # 6. Changing the sentence from "Select who will be authorized to send email from this group list" to "Select who will be authorized to send email to this group list"

The change is needed to ensure that email Administrators understand Group Lists have limitations on who can send email TO them. No user can send FROM a Group List in Rackspace Email based on these Group List settings. Item # 6 was misleading users to believe that this would allow them to send AS the Group List (from the group list) and not TO the Group List. This also lead to users believing that any email address, on any domain, can send email to a Rackspace Email Group List. This is not true. Only users in the domain, and external users that the the administrator specifies, can send emails to Rackspace Email Group Lists.