Update basic security steps for linux
The current version of this article covers the steps to disable root login and an inefficient method of setting up sudo users. This section was removed and a link to the security article for Linux that describes the process in more detail was added. 

Additionally, I don't believe this article should be considered as "basic security". The modifications to the bootloader can cause serious issues if implemented incorrectly and is really more for use if your account is compromised. Alternatively, if the customer just wants an extra layer of security, this can be implemented. 

I know this article was recently provided to customers as part of COVID, however it really should really be changed to indicate that this is advanced security information that is likely not necessary for most customers. Feel free to reach out with any questions.