Fix formatting and add note for filesystem command
Fixed formatting for terminal output on article. 

Also, I added a note box under the command example for clarification.  I think that line might have been in a previous iteration of the article before it was submitted. Figured it couldn't hurt to add the extra info, so I added it to the page.