Create new article - Wordpress FTP (#924)
* Create new article - Wordpress FTP

New article submitted for review. 

Please note that I have included a comment section since I'm not sure on how to best approach a disclaimer. 

Ultimately, Wordpress is not supported at Rackspace outside of an installation and best effort configuration. Any panel level changes need to be handled by the customers as we do not have any access to that environment.

* edits

* edits

* corrections

Co-authored-by: Rose Morales <viridaria@gmail.com>
Co-authored-by: Cat Lookabaugh <cat.lookabaugh@rackspace.com>